+++ src/main/kotlin/com/widehouse/dnd/character/Skill.kt
@@ -1,10 +1,11 @@
 package com.widehouse.dnd.character
 
-import com.widehouse.dnd.character.AbilityType.Charisma
-import com.widehouse.dnd.character.AbilityType.Dexterity
-import com.widehouse.dnd.character.AbilityType.Intelligence
-import com.widehouse.dnd.character.AbilityType.Strength
-import com.widehouse.dnd.character.AbilityType.Wisdom
+import com.widehouse.dnd.character.ability.AbilityType
+import com.widehouse.dnd.character.ability.AbilityType.Charisma
+import com.widehouse.dnd.character.ability.AbilityType.Dexterity
+import com.widehouse.dnd.character.ability.AbilityType.Intelligence
+import com.widehouse.dnd.character.ability.AbilityType.Strength
+import com.widehouse.dnd.character.ability.AbilityType.Wisdom
 
 sealed class Skill(val abilityType: AbilityType)
