+++ src/main/kotlin/com/widehouse/dnd/character/ability/AbilityType.kt
@@ -1,0 +1,10 @@
+package com.widehouse.dnd.character.ability
+
+sealed class AbilityType {
+    object Strength : AbilityType()
+    object Dexterity : AbilityType()
+    object Constitution : AbilityType()
+    object Intelligence : AbilityType()
+    object Wisdom : AbilityType()
+    object Charisma : AbilityType()
+}
