+++ src/main/kotlin/com/widehouse/dnd/character/nonplayer/Monster.kt
@@ -1,6 +1,6 @@
 package com.widehouse.dnd.character.nonplayer
 
-import com.widehouse.dnd.character.Abilities
+import com.widehouse.dnd.character.ability.Abilities
 import com.widehouse.dnd.character.Character
 import com.widehouse.dnd.character.action.Action
 import com.widehouse.dnd.item.Weapon
