+++ src/main/kotlin/com/widehouse/dnd/character/player/Class.kt
@@ -1,12 +1,12 @@
 package com.widehouse.dnd.character.player
 
-import com.widehouse.dnd.character.AbilityType
-import com.widehouse.dnd.character.AbilityType.Charisma
-import com.widehouse.dnd.character.AbilityType.Constitution
-import com.widehouse.dnd.character.AbilityType.Dexterity
-import com.widehouse.dnd.character.AbilityType.Intelligence
-import com.widehouse.dnd.character.AbilityType.Strength
-import com.widehouse.dnd.character.AbilityType.Wisdom
+import com.widehouse.dnd.character.ability.AbilityType
+import com.widehouse.dnd.character.ability.AbilityType.Charisma
+import com.widehouse.dnd.character.ability.AbilityType.Constitution
+import com.widehouse.dnd.character.ability.AbilityType.Dexterity
+import com.widehouse.dnd.character.ability.AbilityType.Intelligence
+import com.widehouse.dnd.character.ability.AbilityType.Strength
+import com.widehouse.dnd.character.ability.AbilityType.Wisdom
 import com.widehouse.dnd.dice.Dice
 
 enum class Class(val hitDice: Dice, val proficiencySavingThrow: List<AbilityType>) {
