+++ src/test/kotlin/com/widehouse/dnd/character/AbilityTest.kt
@@ -1,5 +1,7 @@
 package com.widehouse.dnd.character
 
+import com.widehouse.dnd.character.ability.Dexterity
+import com.widehouse.dnd.character.ability.Strength
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
