+++ src/test/kotlin/com/widehouse/dnd/character/player/PlayerCharacterTest.kt
@@ -1,9 +1,7 @@
 package com.widehouse.dnd.character.player
 
-import com.widehouse.dnd.character.Abilities
-import com.widehouse.dnd.character.AbilityType
-import com.widehouse.dnd.character.Constitution
-import com.widehouse.dnd.character.Strength
+import com.widehouse.dnd.character.ability.Abilities
+import com.widehouse.dnd.character.ability.AbilityType
 import io.kotest.assertions.assertSoftly
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.collections.shouldContain
@@ -11,23 +9,15 @@ import io.kotest.matchers.shouldBe
 
 class PlayerCharacterTest : FreeSpec({
     "Player Character creation test" {
-        val char = PlayerCharacter.create(name = "foo", race = Race.Human, `class` = Class.Fighter, abilities = Abilities.of(
-            15,
-            14,
-            13,
-            12,
-            10,
-            8
-        )
-        )
+        val character = PlayerCharacter.create("foo", Race.Human, Class.Fighter, Abilities.of(15, 14, 13, 12, 10, 8))
 
-        assertSoftly(char) {
+        assertSoftly(character) {
             name shouldBe "foo"
             race shouldBe Race.Human
             `class` shouldBe Class.Fighter
             level shouldBe 1
-            abilities.strength shouldBe Strength(15)
-            abilities.constitution shouldBe Constitution(13)
+            abilities.strength.score shouldBe 15
+            abilities.constitution.score shouldBe 13
             hitPoints shouldBe 11
             proficiencyBonus shouldBe 2
             proficiencySavingThrow shouldContain AbilityType.Strength
