+++ src/testFixtures/kotlin/com/widehouse/dnd/character/CharacterFixtures.kt
@@ -1,5 +1,7 @@
 package com.widehouse.dnd.character
 
+import com.widehouse.dnd.character.ability.Abilities
+
 object CharacterFixtures {
     val foo: Character = Character("foo", Abilities.of(10, 10, 10, 10, 10, 10), 10)
 }
